requirements.txt
lxml==6.0.0
numpy==2.2.6
pandas==2.3.1
python-dateutil==2.9.0.post0
python-docx==1.2.0
pytz==2025.2
six==1.17.0
typing_extensions==4.14.1
tzdata==2025.2
docx2txt==0.8
PyPDF2>=3.0.1
unstructured[all-docs]
pillow
lxml
langchain
langchain-community
langchain-openai
chromadb
tiktoken
python-dotenv
langchain_ollama 
langchain_chroma 
langchain-core
fastapi
requests
folium>=0.15.1
streamlit
streamlit-folium
folium
geopandas 
shapely
osm2geojson
osmnx
pyproj
rasterio